refactor(service-client): move Item to below the Items
https://github.com/akky-xxxx/project-my-homepage/issues/101
