+++ packages/service-client/src/components/templates/PhotoGallery/components/atoms/Items/components/atoms/Item/index.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStoryObj, Meta } from "@storybook/react"
 import { ComponentProps } from "react"
-import { DeviceMap } from "../../../../../../shared/const/DeviceMap"
+import { DeviceMap } from "../../../../../../../../../shared/const/DeviceMap"
 import { Item } from "./index"
 
 type ItemType = typeof Item
